draft3 of #2 backend, might be finished

+++ src/reasonablecare/data/CreditCard.java
@@ -6,7 +6,7 @@ public class CreditCard
    * with a possible credit card check failure
    * the credit check will fail 20% of the time
    */
-  public static boolean checkCardAmount(String ccnum, String ccexp, int amount)
+  public static boolean checkCardAmount(String ccnum, String ccexp, long amount)
   {
     return Math.random() > 0.2;
   }
